Update README.md
Hi I am a student in a technical writing class for computing science (CMPT 376W) and I made some suggestions for your README. I revised the description for "Assignment 3". It would be helpful for your readers to know how to run your application. I happened to do this same assignment for CMPT 300 so I am guessing you are doing the same assignment so I though adding this README would be helpful. Hope you find it useful!